requirements.txt
Django==4.2.10
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
simpy==4.0.2
paho-mqtt==1.6.1
matplotlib==3.7.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0